Move `test_related` to a proper model file
