--- tests/assert_type/db/models/fields/test_related.py
@@ -1,15 +1,0 @@
-from __future__ import annotations
-
-from django.db import models
-
-
-class Author(models.Model):
-    pass
-
-
-class Book(models.Model):
-    author = models.ForeignKey(Author, on_delete=models.CASCADE, swappable=False)
-
-
-class Profile(models.Model):
-    user = models.OneToOneField(Author, on_delete=models.CASCADE, swappable=False)
